BREAKING: merge namespace bcftools to vcf
