+++ tests/test_vcf/Bcftools/data/anno.vcf
